# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: GramSathi AI API
  description: |
    AI-Powered Multilingual Government Assistant API for rural citizens in India.
    Enables scheme discovery, application processing, document OCR, grievance filing,
    notifications, and analytics across 22 Indian languages.
  version: "1.0.0"
  contact:
    name: GramSathi Team
    email: dev@gramsathi.ai
  license:
    name: Proprietary
servers:
  - url: https://api.gramsathi.ai/v1
    description: Production server
  - url: https://staging-api.gramsathi.ai/v1
    description: Staging server
  - url: http://localhost:8001/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token obtained from /auth/otp/verify
  schemas:
...
    GrievanceCreate:
      type: object
      required:
        - citizen_id
        - category
        - subject
        - description
      properties:
        citizen_id:
          type: string
          format: uuid
        category:
          type: string
          enum: [scheme_related, document_issue, payment_delay, staff_behavior, technical, other]
        department:
          type: string
        subject:
          type: string
          maxLength: 200
        description:
          type: string
        original_language:
          type: string
          example: "hi"
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: medium
    GrievanceOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        citizen_id:
          type: string
          format: uuid
        category:
          type: string
        department:
          type: string
          nullable: true
        subject:
          type: string
        description:
          type: string
        original_language:
          type: string
          nullable: true
        priority:
          type: string
        status:
          type: string
          enum: [draft, submitted, under_review, escalated, resolved, closed]
        resolution_notes:
          type: string
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GrievanceUpdate:
      type: object
      properties:
        category:
          type: string
        department:
          type: string
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
    GrievanceDraftRequest:
      type: object
      required:
        - description
      properties:
        description:
          type: string
        category:
          type: string
        original_language:
          type: string
          example: "hi"
    GrievanceDraftResponse:
      type: object
      properties:
        drafted_complaint:
          type: string
        category:
          type: string
        department:
          type: string
        subject:
          type: string
        priority:
          type: string
    ComplaintTrackingOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        grievance_id:
          type: string
          format: uuid
        action:
          type: string
        action_by:
          type: string
          format: uuid
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    # Notification Schemas
...
paths:
...
  /grievances:
    post:
      tags:
        - Grievances
      summary: Create a new grievance
      operationId: createGrievance
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrievanceCreate"
            example:
              citizen_id: "550e8400-e29b-41d4-a716-446655440000"
              category: "scheme_related"
              department: "Department of Social Welfare"
              subject: "PM Awas Yojana application pending for 6 months"
              description: "I applied for PM Awas Yojana in January but have not received any update."
              original_language: "hi"
              priority: "high"
      responses:
        "201":
          description: Grievance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrievanceOut"
  /grievances/pending:
    get:
      tags:
        - Grievances
      summary: List pending grievances (gov officer)
      operationId: listPendingGrievances
      security:
        - BearerAuth: []
      parameters:
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: page_size
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Pending grievances list
  /grievances/{grievance_id}:
    get:
      tags:
        - Grievances
      summary: Get grievance details
      operationId: getGrievance
      security:
        - BearerAuth: []
      parameters:
        - name: grievance_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Grievance details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrievanceOut"
    put:
      tags:
        - Grievances
      summary: Update grievance
      operationId: updateGrievance
      security:
        - BearerAuth: []
      parameters:
        - name: grievance_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrievanceUpdate"
      responses:
        "200":
          description: Grievance updated
  /grievances/{grievance_id}/submit:
    post:
      tags:
        - Grievances
      summary: Submit grievance for processing
      operationId: submitGrievance
      security:
        - BearerAuth: []
      parameters:
        - name: grievance_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Grievance submitted
  /grievances/{grievance_id}/escalate:
    post:
      tags:
        - Grievances
      summary: Escalate grievance (gov officer)
      operationId: escalateGrievance
      security:
        - BearerAuth: []
      parameters:
        - name: grievance_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Grievance escalated
  /grievances/{grievance_id}/resolve:
    post:
      tags:
        - Grievances
      summary: Resolve grievance (gov officer)
      operationId: resolveGrievance
      security:
        - BearerAuth: []
      parameters:
        - name: grievance_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - resolution_notes
              properties:
                resolution_notes:
                  type: string
      responses:
        "200":
          description: Grievance resolved
  /grievances/{grievance_id}/reopen:
    post:
      tags:
        - Grievances
      summary: Reopen resolved grievance
      operationId: reopenGrievance
      security:
        - BearerAuth: []
      parameters:
        - name: grievance_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Grievance reopened
  /grievances/{grievance_id}/tracking:
    get:
      tags:
        - Grievances
      summary: Get grievance tracking history
      operationId: getGrievanceTracking
      security:
        - BearerAuth: []
      parameters:
        - name: grievance_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tracking history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplaintTrackingOut"
  /citizens/{citizen_id}/grievances:
    get:
      tags:
        - Grievances
      summary: List citizen's grievances
      operationId: listCitizenGrievances
      security:
        - BearerAuth: []
      parameters:
        - name: citizen_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: page_size
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Citizen's grievances
  /ai/grievances/draft:
    post:
      tags:
        - Grievances
      summary: AI-assisted grievance drafting
      operationId: aiDraftGrievance
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrievanceDraftRequest"
      responses:
        "200":
          description: AI-drafted grievance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrievanceDraftResponse"

  # ===== Notification Service =====
...